Move Contributor classes to a module
This refactor doesn't make any code changes, just moves the
Contributor class and its two related Name and Role classes into
separate files in a module for easier reading.
